Update dependencies to newer versions. Some of the test cases are broken, due to private accessor. Update Docker to use Ubuntu 18.04

+++ examples/counter_service/counter_router.cpp
@@ -26,7 +26,7 @@ uint32_t ShardId(const std::string& key, const uint32_t num_shards) {
     hash_code = 31 * hash_code + key[i];
   }
 
-  return abs(hash_code % num_shards);
+  return abs((int)(hash_code % num_shards));
 }
 
 }  // anonymous namespace
